:art: refactor : 클래스 명 개선
